requirements.txt
sentence-transformers
faiss-cpu
inference
flask
flask-cors
opencv-python
nltk
groq